requirement.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==15.1.1
filelock==3.6.0
mysql-connector-python==8.0.28
Pillow==9.0.1
platformdirs==2.5.1
PyJWT==2.5.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.3
virtualenvwrapper-win==1.2.7
